Lower network usage when sending whitelist data
This uses about a 6th of the data we were using before, might be a bit
more. Only issue with this is it seems to create a bit more client-side
lag, the shitty little laptop I'm testing all this on lags for about 1
second on whitelist with 500 entries. Hopefully I can come up with a
better way to do this at some point.